+++ resources/js/components/Role/RolesComponent.vue
@@ -26,7 +26,12 @@
 
       <template v-slot:row-details="role">
         <div class="d-flex justify-content-end">
-          <b-button variant="outline-info" class="mr-2" @click="showEditForm(role.item)" v-if="canUpdate">Edit</b-button>
+          <b-button
+            variant="outline-info"
+            class="mr-2"
+            @click="showEditForm(role.item)"
+            v-if="canUpdate"
+          >Edit</b-button>
           <b-button variant="outline-danger" @click="destroy(role.item.id)" v-if="canDelete">Delete</b-button>
         </div>
       </template>
@@ -35,16 +40,21 @@
       v-model="currentPage"
       :total-rows="rows"
       :per-page="perPage"
-      aria-controls=""
+      aria-controls
       first-number
       last-number
     ></b-pagination>
-    <role-form-component v-if="canUpdate || canCreate" @createRole="createRole" :role="selectedRole" @updateRole="updateRole"></role-form-component>
+    <role-form-component
+      v-if="canUpdate || canCreate"
+      @createRole="createRole"
+      :role="selectedRole"
+      @updateRole="updateRole"
+    ></role-form-component>
   </div>
 </template>
 <script>
 import roleDataService from "../../services/role-data-service";
-import toastOptions from "../../services/toast-options";
+import { toastOptions, msgBoxOptions } from "../../services/toast-options";
 
 export default {
   data() {
@@ -56,7 +66,7 @@ export default {
           key: "name",
           sortable: true,
           label: "Name",
-        }
+        },
       ],
       roles: [],
       perPage: 50,
@@ -76,19 +86,34 @@ export default {
       this.$bvModal.show("role-form-modal");
     },
     destroy(id) {
-      this.loadingData = true;
-      roleDataService.delete(id).then((response) => {
-        if (response.data.id == id) {
-          this.roles.splice(
-            this.roles.findIndex((x) => x.id == id),
-            1
-          );
-          this.$bvToast.toast("Role was deleted successfully!", toastOptions());
-        } else {
-          this.$bvToast.toast("Something went wrong", toastOptions("danger"));
-        }
-        this.loadingData = false;
-      });
+      this.$bvModal
+        .msgBoxConfirm(
+          "Are you sure you want to delete this role?",
+          msgBoxOptions()
+        )
+        .then((value) => {
+          if (value == true) {
+            this.loadingData = true;
+            roleDataService.delete(id).then((response) => {
+              if (response.data.id == id) {
+                this.roles.splice(
+                  this.roles.findIndex((x) => x.id == id),
+                  1
+                );
+                this.$bvToast.toast(
+                  "Role was deleted successfully!",
+                  toastOptions()
+                );
+              } else {
+                this.$bvToast.toast(
+                  "Something went wrong",
+                  toastOptions("danger")
+                );
+              }
+              this.loadingData = false;
+            });
+          }
+        });
     },
     updateRole(role) {
       this.roles.splice(
@@ -114,14 +139,14 @@ export default {
       return this.roles.length;
     },
     canCreate() {
-      return this.$can('role.create');
+      return this.$can("role.create");
     },
     canUpdate() {
-      return this.$can('role.update');
+      return this.$can("role.update");
     },
     canDelete() {
-      return this.$can('role.delete');
-    }
+      return this.$can("role.delete");
+    },
   },
 };
 </script>
